Move database utility to separate file
